+++ lib/qfp.tsx
